Chapter 7.2 and 7.3 finished
Both are very important chapters talking about 18th century. Here it is about Nationalism, the doctrine to have a state of your own people and modern imperialism and the race between countries at this time.

Note that you need to practice cause and effect with history and so it would be much use for you to try and understand how cause and effect leads to major historical changes.